feat: remove relation requirement from assign

Remove relation requirement from validate


fix remove policies


feat: remove policy check


fix: ci


fix: tests


fix tests


fix api docs


fix: property based tests


revert add policies


Refactor relation check


fix: ci

+++ auth/postgres/domains.go
@@ -370,6 +370,17 @@ func (repo domainRepo) DeletePolicies(ctx context.Context, pcs ...auth.Policy) (
 
 	for _, pc := range pcs {
 		q := `
+		DELETE FROM
+			policies
+		WHERE
+			subject_type = :subject_type
+			AND subject_id = :subject_id
+			AND subject_relation = :subject_relation
+			AND object_type = :object_type
+			AND object_id = :object_id
+		;`
+		if pc.Relation != "" {
+			q = `
 			DELETE FROM
 				policies
 			WHERE
@@ -380,6 +391,7 @@ func (repo domainRepo) DeletePolicies(ctx context.Context, pcs ...auth.Policy) (
 				AND object_type = :object_type
 				AND object_id = :object_id
 			;`
+		}
 
 		dbpc := toDBPolicy(pc)
 		row, err := tx.NamedQuery(q, dbpc)
